# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateDefinitionRequest:
    properties:
      category:
        maxLength: 50
        type: string
      description:
        maxLength: 2000
        type: string
      key:
        maxLength: 100
        type: string
      name:
        maxLength: 200
        type: string
    required:
    - key
    - name
    type: object
...
  dto.DefinitionResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      draft_graph:
        $ref: '#/definitions/dto.GraphData'
      id:
        type: string
      key:
        type: string
      name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  dto.GraphData:
    properties:
      edges:
        items:
          $ref: '#/definitions/dto.GraphEdge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/dto.GraphNode'
        type: array
    type: object
  dto.GraphEdge:
    properties:
      data:
        additionalProperties: true
        type: object
      id:
        type: string
      label:
        type: string
      source:
        type: string
      sourceHandle:
        type: string
      target:
        type: string
    type: object
  dto.GraphNode:
    properties:
      data:
        additionalProperties: true
        type: object
      id:
        type: string
      position:
        $ref: '#/definitions/dto.GraphPosition'
      type:
        type: string
    type: object
  dto.GraphPosition:
    properties:
      x:
        type: number
      "y":
        type: number
    type: object
...
  dto.PublishDefinitionRequest:
    properties:
      graph_data:
        $ref: '#/definitions/dto.GraphData'
    required:
    - graph_data
    type: object
...
  dto.SaveDraftRequest:
    properties:
      graph_data:
        $ref: '#/definitions/dto.GraphData'
    required:
    - graph_data
    type: object
...
  dto.UpdateDefinitionRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      name:
        maxLength: 200
        type: string
    required:
    - name
    type: object
...
  dto.VersionResponse:
    properties:
      bpmn_data:
        $ref: '#/definitions/dto.GraphData'
      created_at:
        type: string
      definition_id:
        type: string
      id:
        type: string
      published_at:
        type: string
      published_by:
        type: string
      status:
        type: integer
      version:
        type: integer
    type: object
...
  response.PageResponse:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 高校计算机协会一体化管理平台
  title: StarByte API
  version: "1.0"
paths:
...
  /workflow/definitions:
    get:
      consumes:
      - application/json
      description: 分页查询流程定义列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 分类
        in: query
        name: category
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResponse'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/dto.DefinitionResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取流程定义列表
      tags:
      - 流程定义
    post:
      consumes:
      - application/json
      description: 创建新的流程定义（草稿状态）
      parameters:
      - description: 流程定义信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DefinitionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建流程定义
      tags:
      - 流程定义
  /workflow/definitions/{id}:
    delete:
      description: 删除指定流程定义
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除流程定义
      tags:
      - 流程定义
    get:
      description: 根据ID获取流程定义详细信息
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DefinitionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取流程定义详情
      tags:
      - 流程定义
    put:
      consumes:
      - application/json
      description: 更新流程定义基本信息
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      - description: 流程定义信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DefinitionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新流程定义
      tags:
      - 流程定义
  /workflow/definitions/{id}/draft:
    put:
      consumes:
      - application/json
      description: 保存未发布的工作稿 graph_data，已发布定义也可保存下一版草稿
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      - description: 草稿流程图数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SaveDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DefinitionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 保存流程定义草稿图
      tags:
      - 流程定义
  /workflow/definitions/{id}/publish:
    post:
      consumes:
      - application/json
      description: 发布流程定义，创建新版本
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      - description: 流程图数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PublishDefinitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VersionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 发布流程定义
      tags:
      - 流程定义
  /workflow/definitions/{id}/versions:
    get:
      description: 获取指定流程定义的所有版本
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VersionResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取流程定义版本列表
      tags:
      - 流程定义
  /workflow/definitions/{id}/versions/{versionId}:
    get:
      description: 根据版本ID获取流程定义版本详情，并校验版本归属
      parameters:
      - description: 流程定义ID
        in: path
        name: id
        required: true
        type: string
      - description: 版本ID
        in: path
        name: versionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.VersionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取指定版本详情
      tags:
      - 流程定义
...
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"